requirements.txt
google-genai
mcp
python-dotenv
